+++ src/Promitor.Scraper.Host/Validation/Steps/MetricsDeclarationValidationStep.cs
@@ -6,6 +6,7 @@ using Promitor.Scraper.Host.Configuration.Model.Metrics;
 using Promitor.Scraper.Host.Configuration.Model.Metrics.ResouceTypes;
 using Promitor.Scraper.Host.Configuration.Providers.Interfaces;
 using Promitor.Scraper.Host.Validation.Interfaces;
+using Promitor.Scraper.Host.Validation.MetricDefinitions;
 
 namespace Promitor.Scraper.Host.Validation.Steps
 {
@@ -80,77 +81,6 @@ namespace Promitor.Scraper.Host.Validation.Steps
             return errorMessages;
         }
 
-        private List<string> ValidateAzureMetricConfiguration(AzureMetricConfiguration azureMetricConfiguration)
-        {
-            var errorMessages = new List<string>();
-
-            if (azureMetricConfiguration == null)
-            {
-                errorMessages.Add("Invalid azure metric configuration is configured");
-                return errorMessages;
-            }
-
-            if (string.IsNullOrWhiteSpace(azureMetricConfiguration.MetricName))
-            {
-                errorMessages.Add("No metric name for Azure is configured");
-            }
-
-            return errorMessages;
-        }
-
-        private List<string> ValidateMetric(MetricDefinition metric)
-        {
-            var errorMessages = new List<string>();
-
-            if (metric == null)
-            {
-                errorMessages.Add("Invalid metric is configured");
-                return errorMessages;
-            }
-
-            if (metric.ResourceType == ResourceType.NotSpecified)
-            {
-                errorMessages.Add($"{metric.ResourceType} '{nameof(ResourceType.NotSpecified)}' is not supported");
-            }
-
-            if (string.IsNullOrWhiteSpace(metric.Name))
-            {
-                errorMessages.Add("No metric name is configured");
-            }
-
-            switch (metric.ResourceType)
-            {
-                case ResourceType.ServiceBusQueue:
-                    ValidateServiceBusQueueMetricConfiguration(metric as ServiceBusQueueMetricDefinition, errorMessages);
-                    break;
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(metric), metric.ResourceType, $"No validation rules are defined for metric type '{metric.ResourceType}'");
-            }
-
-            var metricsConfigurationErrorMessages = ValidateAzureMetricConfiguration(metric.AzureMetricConfiguration);
-            errorMessages.AddRange(metricsConfigurationErrorMessages);
-
-            return errorMessages;
-        }
-
-        private void ValidateServiceBusQueueMetricConfiguration(ServiceBusQueueMetricDefinition serviceBusQueueMetricDefinition, List<string> errorMessages)
-        {
-            if (serviceBusQueueMetricDefinition == null)
-            {
-                return;
-            }
-
-            if (string.IsNullOrWhiteSpace(serviceBusQueueMetricDefinition.Namespace))
-            {
-                errorMessages.Add("No Service Bus Namespace is configured");
-            }
-
-            if (string.IsNullOrWhiteSpace(serviceBusQueueMetricDefinition.QueueName))
-            {
-                errorMessages.Add("No queue name is configured");
-            }
-        }
-
         private List<string> ValidateMetrics(List<MetricDefinition> metrics)
         {
             var errorMessages = new List<string>();
@@ -161,11 +91,9 @@ namespace Promitor.Scraper.Host.Validation.Steps
                 return errorMessages;
             }
 
-            foreach (var metric in metrics)
-            {
-                var metricErrorMessages = ValidateMetric(metric);
-                errorMessages.AddRange(metricErrorMessages);
-            }
+            var metricsValidator = new MetricsValidator();
+            var metricErrorMessages = metricsValidator.Validate(metrics);
+            errorMessages.AddRange(metricErrorMessages);
 
             // Detect duplicate metric names
             var duplicateMetricNames = DetectDuplicateMetrics(metrics);
